+++ packages/extensions/src/Plugins/Youtube/index.ts
@@ -4,6 +4,7 @@ import { getSongList } from './utils';
 
 export function getAudioUrl(youtubeUrl: string) {
     console.log(youtubeUrl);
+    console.log(ytdl.validateURL(youtubeUrl));
     return ytdl(youtubeUrl, { filter: format => format.container === 'mp4' })
         .then(urls => {
             const { url } = urls[0];